refactoring paquetes parte 2.
