+++ Emisor/Client/src/modelo_e/ArmableDesarmable.java
@@ -1,4 +1,4 @@
-package modelo;
+package modelo_e;
 
 public interface ArmableDesarmable {
     static final String SEPARADOR = "-"; // capaz en vez de leer por linea cambiar el separador y leer todo junto
