+++ Emisor/Client/src/modelo_e/RearmadorLista.java
@@ -1,4 +1,4 @@
-package modelo;
+package modelo_e;
 
 import java.util.ArrayList;
