+++ Emisor/Client/src/modelo_e/ServerRecepcion.java
@@ -1,4 +1,4 @@
-package modelo;
+package modelo_e;
 
 import java.io.BufferedReader;
 import java.io.IOException;
