+++ Emisor/Client/src/modelo_e/mensaje/MensajeRecepcion.java
@@ -1,9 +1,9 @@
-package modelo.mensaje;
+package modelo_e.mensaje;
 
 import excepciones.MensajeMalFormadoException;
 
-import modelo.ArmableDesarmable;
-import modelo.agenda.Usuario;
+import modelo_e.ArmableDesarmable;
+import modelo_e.agenda.Usuario;
 
 public class MensajeRecepcion extends Mensaje {
     public MensajeRecepcion(String asunto, String descripcion, Usuario destinatario, Usuario emisor) {
